backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-docx==1.1.2
docxtpl==0.16.8
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
psycopg2-binary==2.9.9
aiofiles==23.2.1
Pillow==10.3.0
